11st of June, source code of chat is updated clearly.
