+++ src/main/java/com/spring/fleamarket/domain/chat/controller/APIController.java
@@ -6,21 +6,34 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.spring.fleamarket.domain.account.service.AccountFindService;
 import com.spring.fleamarket.domain.chat.model.MessageModel;
 import com.spring.fleamarket.domain.chat.service.MessageManageService;
+import com.spring.fleamarket.domain.chat.service.ReportService;
 import com.spring.fleamarket.domain.model.Account;
-import com.spring.fleamarket.global.security.annotation.LoginedAccount;
+import com.spring.fleamarket.domain.model.Report;
 
+import lombok.extern.log4j.Log4j;
+
+@Log4j
 @RestController
 @CrossOrigin
-public class UsersController {
+public class APIController {
 	
 	@Autowired
 	MessageManageService service;
+	
+	@Autowired
+	AccountFindService serviceAF;
+	
+	@Autowired
+	ReportService serviceRP;
 	
 	@GetMapping("/recallChat/{sender}/{receiver}")
 	public List<MessageModel> bringUpMessageRecord(MessageModel msgModel, @PathVariable int sender, @PathVariable int receiver) {
@@ -31,6 +44,24 @@ public class UsersController {
 			
 		return service.recallChat(msgModel);
 		
+	}
+	
+	@GetMapping("/chat")
+	public List<Account> bringUpChatUser() {
+		
+		//현재의 아이디와 채팅한 유저만 불어오기(수정필요)
+		List<Account> result = serviceAF.selectAll(); 
+	
+		return result;
+	}
+	
+	@ResponseBody
+	@RequestMapping(value="/report", method = RequestMethod.POST)
+	public void saveReportInfo(@RequestBody Report rp) {
+			
+		log.info(rp.toString());
+	
+		serviceRP.saveReport(rp);
 	}
