+++ src/main/java/com/spring/fleamarket/domain/chat/mapper/MessageManageMapper.java
@@ -11,6 +11,6 @@ public interface MessageManageMapper {
 	
 	public List<MessageModel> recallChat(MessageModel msgModel);
 	
-	public Date checkTime(String content);
+	public Date checkTime(int messageNo);
 
 }
