+++ src/main/resources/com/spring/fleamarket/domain/chat/mapper/MessageManageMapper.xml
@@ -2,9 +2,10 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.spring.fleamarket.domain.chat.mapper.MessageSaveMapper">
+<mapper namespace="com.spring.fleamarket.domain.chat.mapper.MessageManageMapper">
 
-	<insert id="saveMessage">
+	<insert id="saveMessage" useGeneratedKeys="true"
+    keyProperty="messageNo">
 		INSERT INTO
 			messages(sender_id, receiver_id, content)
 		VALUES
@@ -20,10 +21,8 @@
 	<select id="checkTime" resultType="java.util.Date">
 		SELECT create_date
 		FROM messages
-		WHERE content = #{content}
-		ORDER BY create_date
-		DESC LIMIT 1
-	
+		WHERE message_no = #{messageNo}
+
 	</select>
 	
 </mapper>
